chore: make build_subset more modular by moving the to_bool func to our conversion utils module
