+++ src/utils/__init__.py
@@ -1,4 +1,4 @@
-from .conversion_utils import generate_case_variations
+from .conversion_utils import generate_case_variations, to_bool
 from .data_dict_utils import get_true_false_from_data_dict
 from .file_utils import load_zstd_to_batches
 from .schema_utils import load_schema_from_json
@@ -6,6 +6,7 @@ from .schema_utils import load_schema_from_json
 __all__ = [
     "generate_case_variations",
     "get_true_false_from_data_dict",
+    "to_bool",
     "load_zstd_to_batches",
     "load_schema_from_json",
 ]
